svn merge -r 19417:20113 https://dart.googlecode.com/svn/branches/bleeding_edge trunk

git-svn-id: http://dart.googlecode.com/svn/trunk@20121 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/lib/date_patch.dart
@@ -183,7 +183,7 @@ patch class DateTime {
     // Simplify calculations by working with zero-based month.
     --month;
     // Deal with under and overflow.
-    year += (month / 12).floor().toInt();
+    year += (month / 12).floor();
     month = month % 12;
 
     // First compute the seconds in UTC, independent of the [isUtc] flag. If
